+++ src/compat/src/volume/tests/mod.rs
@@ -1,5 +1,0 @@
-mod filesystem;
-mod mount;
-mod repository;
-mod service;
-pub(crate) mod support;